Standardize Assert imports
The project uses a mixed convention of importing org.junit.Assert and
importing static methods from that class, sometimes even in the same
class.
This isn't a bug per-se, but can be confusing, and makes refactoring a
bit harder.

This patch standardizes the entire project to use the de-facto
standard set by most of the code of importing static methods from
org.junit.Assert.