Fix CompositeCronTest
CompositeCronTest's weDoNotSupportCronsWithDifferentDefinitions
doesn't fail, but it is subtly wrong.

The statement "parser2.parse("0 0 1 * *")" throws an
IllegalArgumentException (causing the test to pass), and by doing so
the call to the CompositeCron's constructor which is what this method
is supposed to test is never called.

This patch fixes the initialization of the cron2 variable to a valid
CronParser so that the CompositeCron constructor may be called, and
the test can prove it throws the expected IllegalArgumentException.