Fix BetweenFieldValueGeneratorTest
BetweenFieldValueGeneratorTest's testGenerateNextValue and
testGeneratePreviousValue don't fail, but they are subtly wrong.

In both tests, the loops have one iteration too many, meaning the last
iteration of the loop will throw the expected NoSuchValueException,
and the statement after the loop will never be executed.

This patch fixes the test to achieve the intended functionally - for
each iteration of the loop the call to the tested method should
succeed and the assertion should pass. Once the loop is finished, the
extra call after the loop should throw the expected
NoSuchValueException.